+++ src/alsyntaxmodifiers/iRemoveUnusedVariablesSettings.ts
@@ -1,0 +1,5 @@
+export interface IRemoveUnusedVariablesSettings {
+    removeGlobalVariables: boolean | undefined;
+    removeLocalVariables: boolean | undefined;
+    removeLocalMethodParameters: boolean | undefined;
+}