+++ src/alsyntaxmodifiers/removeUnusedVariablesModifier.ts
@@ -1,31 +1,56 @@
 import * as vscode from 'vscode';
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
 import { NameValueQuickPickItem } from '../tools/nameValueQuickPickItem';
+import { IRemoveUnusedVariablesSettings } from './iRemoveUnusedVariablesSettings';
 import { WorkspaceCommandSyntaxModifier } from './workspaceCommandSyntaxModifier';
 
 export class RemoveUnusedVariablesModifier extends WorkspaceCommandSyntaxModifier {
-    protected _variableTypes: any;
+    protected _variableTypes: IRemoveUnusedVariablesSettings;
 
     constructor(context: DevToolsExtensionContext) {
         super(context, "Remove Unused Variables", "removeUnusedVariables");
-        this._variableTypes = {};
+        this._variableTypes = {
+            removeGlobalVariables: false,
+            removeLocalVariables: false,
+            removeLocalMethodParameters: false
+        };
     }
 
     protected getParameters(uri: vscode.Uri): any {
         let parameters = super.getParameters(uri);
-        parameters.removeGlobalVariables = this._variableTypes.removeGlobalVariables;
-        parameters.removeLocalVariables = this._variableTypes.removeLocalVariables;
-        parameters.removeLocalMethodParameters = this._variableTypes.removeLocalMethodParameters;
-
+        this.copySettings(parameters, this._variableTypes);
         return parameters;
+    }
+
+    protected areSettingsEmpty(value: IRemoveUnusedVariablesSettings | undefined) {
+        return ((!value) || (
+            (!value.removeGlobalVariables) &&
+            (!value.removeLocalVariables) &&
+            (!value.removeLocalMethodParameters)));
+    }
+
+    protected copySettings(dest: any, src: any) {
+        if (!src)
+            src = {};
+        dest.removeGlobalVariables = !!src.removeGlobalVariables;
+        dest.removeLocalVariables = !!src.removeLocalVariables;
+        dest.removeLocalMethodParameters = !!src.removeLocalMethodParameters;
+    }
+
+    protected loadDefaultParameters(uri: vscode.Uri | undefined): boolean {
+        let defaultParameters = vscode.workspace.getConfiguration('alOutline', uri).get<IRemoveUnusedVariablesSettings>('defaultRemoveUnusedVariablesSettings');
+        if (this.areSettingsEmpty(defaultParameters))
+            return false;
+        this.copySettings(this._variableTypes, defaultParameters);
+        return true;
     }
 
     async askForParameters(uri: vscode.Uri | undefined): Promise<boolean> {
         this.loadState();
         let quickPickItems = [
-            new NameValueQuickPickItem('Global variables', 'removeGlobalVariables', this._variableTypes.removeGlobalVariables),
-            new NameValueQuickPickItem('Local variables', 'removeLocalVariables', this._variableTypes.removeLocalVariables),
-            new NameValueQuickPickItem('Local methods parameters', 'removeLocalMethodParameters', this._variableTypes.removeLocalMethodParameters)
+            new NameValueQuickPickItem('Global variables', 'removeGlobalVariables', !!this._variableTypes.removeGlobalVariables),
+            new NameValueQuickPickItem('Local variables', 'removeLocalVariables', !!this._variableTypes.removeLocalVariables),
+            new NameValueQuickPickItem('Local methods parameters', 'removeLocalMethodParameters', !!this._variableTypes.removeLocalMethodParameters)
         ];
 
         let selectedValues = await vscode.window.showQuickPick(
@@ -35,18 +60,22 @@ export class RemoveUnusedVariablesModifier extends WorkspaceCommandSyntaxModifie
             return false;
 
         this.clearVariableTypes();
+        let data: any = {};
         if (selectedValues) {
             for (let i=0; i<selectedValues.length; i++) {
-                this._variableTypes[selectedValues[i].value] = true;
+                data[selectedValues[i].value] = true;
             }
         }
+        if (this.areSettingsEmpty(data))
+            return false;
+            
+        this.copySettings(this._variableTypes, data);
         this.saveState();
 
         return true;
     }
 
     private loadState() {
-        this._variableTypes = {};
         let vsctx = this._context.vscodeExtensionContext;
         this._variableTypes.removeGlobalVariables = !!vsctx.globalState.get<boolean>("azALDevTools.remUVar.removeGlobalVariables");
         this._variableTypes.removeLocalVariables = !!vsctx.globalState.get<boolean>("azALDevTools.remUVar.removeLocalVariables");
@@ -64,9 +93,9 @@ export class RemoveUnusedVariablesModifier extends WorkspaceCommandSyntaxModifie
 
     private saveState() {
         let vsctx = this._context.vscodeExtensionContext;
-        vsctx.globalState.update("azALDevTools.remUVar.removeGlobalVariables", this._variableTypes.removeGlobalVariables);
-        vsctx.globalState.update("azALDevTools.remUVar.removeLocalVariables", this._variableTypes.removeLocalVariables);
-        vsctx.globalState.update("azALDevTools.remUVar.removeLocalMethodParameters", this._variableTypes.removeLocalMethodParameters);
+        vsctx.globalState.update("azALDevTools.remUVar.removeGlobalVariables", !!this._variableTypes.removeGlobalVariables);
+        vsctx.globalState.update("azALDevTools.remUVar.removeLocalVariables", !!this._variableTypes.removeLocalVariables);
+        vsctx.globalState.update("azALDevTools.remUVar.removeLocalMethodParameters", !!this._variableTypes.removeLocalMethodParameters);
     }
 
     private clearVariableTypes() {