+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## 3.0.26
  - Issue #310 - SortProperties fails in case of extra semicolon
+ - Issue #311 - Run code cleanup - specify default parameters
 
 Thank you
- - fvet for reporting issue #310
+ - fvet for reporting issues #310 and #311
 
 ## 3.0.25
  - Issue #249 - Reuse tooltip from other pages